src/recommendationservice/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
grpcio==1.32.0            # via -r requirements.in, opentelemetry-instrumentation-grpc
newrelic_telemetry_sdk==0.4.1  # via -r requirements.in
opentelemetry-api==0.14b0  # via opentelemetry-instrumentation, opentelemetry-instrumentation-grpc, opentelemetry-sdk
opentelemetry-instrumentation-grpc==0.14b0  # via -r requirements.in
opentelemetry-instrumentation==0.14b0  # via -r requirements.in
opentelemetry-sdk==0.14b0  # via -r requirements.in, opentelemetry-instrumentation, opentelemetry-instrumentation-grpc
protobuf==3.13.0          # via -r requirements.in
python-json-logger==2.0.1  # via -r requirements.in
six==1.12.0               # via grpcio, protobuf
urllib3==1.25.10          # via newrelic-telemetry-sdk
wrapt==1.12.1             # via opentelemetry-instrumentation

# The following packages are considered to be unsafe in a requirements file:
# setuptools
